Rag-Techniques/hybrid-rag-cag/requirements.txt
sentence-transformers>=2.2.2
faiss-cpu>=1.7.4
redis>=4.5.5
transformers>=4.34.0
datasets>=2.13.0
tqdm>=4.65.0
fastapi>=0.95.0
uvicorn>=0.22.0
python-dotenv>=1.0.0
openai>=1.8.0
pydantic>=1.10.7
gunicorn>=20.1.0
pytest>=7.0.0
fakeredis>=2.17.0